feat(async): Migrating to an async function api for all class methods and the hasher. No more stream
At the end of the day this is all just much simpler for all implementers. Streams should probably
be build on top of this spec so that a chunker is used to hash each relevant portion of the file.

BREAKING CHANGE: MAJOR API BREAK OF EVERY METHOD AND THE HASHER.